+++ config/api/resource-type/summary-parameters.php
@@ -3,14 +3,6 @@
 declare(strict_types=1);
 
 return [
-    'collection' => [
-        'include_resources' => [
-            'field' => 'include-resources',
-            'title' => 'resource-type/parameters.title-include-resources',
-            'description' => 'resource-type/parameters.description-include-resources',
-            'type' => 'boolean',
-            'required' => false
-        ]
-    ],
+    'collection' => [],
     'item' => []
 ];