+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 The complete changelog for the Costs to Expect REST API, follows the format defined at https://keepachangelog.com/en/1.0.0/
 
 ## [v1.14.0] - 2019-05-xx
+### Added
+- New route /summary/resource-types.
+
 ### Changed
 - Content updates, README, CHANGELOG and the landing page.
 - GET parameter changed to include-resources for resource-types route.